Increase max rows to display
Intended to address https://github.com/resurrecting-open-source-projects/cbm/issues/2
about only showing 7 interfaces.

The max number of rows was just a "magic number" set to 8 with a //TODO
next to it. I've changed it to a #define and set it to 64. Admittedly
this is also arbitrary, but it's less likely to cut off useful data
than 8.